third_party/WebKit/LayoutTests/webaudio/doppler-deprecated-expected.txt

CONSOLE WARNING: line 1: 'AudioListener.dopplerFactor' is deprecated and will be removed in M55, around November 2016. It has no effect as the Web Audio doppler effects have already been removed internally. See https://www.chromestatus.com/features/5238926818148352 for more details.
CONSOLE WARNING: line 1: 'AudioListener.speedOfSound' is deprecated and will be removed in M55, around November 2016. It has no effect as the Web Audio doppler effects have already been removed internally. See https://www.chromestatus.com/features/5238926818148352 for more details.
CONSOLE WARNING: line 1: 'AudioListener.setVelocity()' is deprecated and will be removed in M55, around November 2016. It has no effect as the Web Audio doppler effects have already been removed internally. See https://www.chromestatus.com/features/5238926818148352 for more details.
CONSOLE WARNING: line 1: 'PannerNode.setVelocity()' is deprecated and will be removed in M55, around November 2016. It has no effect as the Web Audio doppler effects have already been removed internally. See https://www.chromestatus.com/features/5238926818148352 for more details.
Tests doppler deprecation messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS listener.dopplerFactor is 1
listener.dopplerFactor = 2
PASS listener.dopplerFactor is 2
PASS listener.speedOfSound is 343.3
listener.speedOfSound = 400
PASS listener.speedOfSound is 400
listener.setVelocity(0, 0, 0)
panner.setVelocity(0, 0, 0)
PASS successfullyParsed is true

TEST COMPLETE
